+++ wordseg/statistics.py
@@ -1,75 +1,154 @@
-"""Extract statistics about gold or segmented text"""
+# coding: utf-8
+
+"""Extract statistics relevant for word segmenation corpora"""
+
+# From https://github.com/alecristia/CDSwordSeg/blob/master/recipes/CatalanSpanish/_describe_gold.sh
+# and https://github.com/ConstantineLignos/WordSegmentation.
+# Copyright (C) 2010, 2011 Constantine Lignos
 
 import collections
+import itertools
 import pandas as pd
+
+from math import log2
 
 from wordseg import utils
 from wordseg.separator import Separator
 
 
-default_separator = Separator(phone=None, syllable=None, word=' ')
-"""Default separator when processing gold text
+class CorpusStatistics(object):
+    def __init__(self, corpus, separator, log=utils.null_logger()):
+        self.separator = separator
+        if not self.separator.word:
+            raise ValueError('word separator not defined')
+        if not self.separator.phone:
+            log.warning('phone separator not defined, some stats ignored')
+        log.info('token separator is: %s', self.separator)
 
-Only word tokens, separated by spaces.
+        # force to list and ignore empty lines
+        self.corpus = list(utt.strip() for utt in corpus if len(utt.strip()))
+        log.info('loaded %s utterances', len(self.corpus))
 
-"""
+        # tokenize the text at each defined level ('word', 'syllable'
+        # and/or 'phone') TODO can be optimize we are tokenizing the
+        # entire text up to 3 times
+        self.tokens = {}
+        for level in self.separator.levels():
+            self.tokens[level] = [
+                list(self.separator.tokenize(utt, level, keep_boundaries=False))
+                for utt in self.corpus]
+
+        # estimates token frequencies at each level
+        self.unigram = self._unigram('word')
+
+    def _unigram(self, level):
+        """Return dictionary of (token: frequency) items"""
+        count =  collections.Counter(
+            (t for utt in self.tokens[level] for t in utt)).most_common()
+        total_count = sum(c[1] for c in count)
+        return {c[0]: c[1] / total_count for c in count}
+
+    def describe_corpus(self):
+        """Basic description of the corpus
+
+        Python implementation of https://github.com/alecristia/
+        CDSwordSeg/blob/master/recipes/CatalanSpanish/_describe_gold.sh
+
+        Raises
+        ------
+        If 'phone' and 'word' tokens are not definide in the separator.
+
+        """
+        # length of utterances in number of words
+        wlen = [len(utt) for utt in self.tokens['word']]
+
+        # the list of all words in the corpus
+        words = [w for u in self.tokens['word'] for w in u]
+
+        # ratio of uniques words per chunk of 10 words
+        nuniques = [
+            len(set(words[x:x+10])) / 10 for x in range(len(words) - 10)]
+
+        stats = {
+            # number of utterances
+            'nutterances': len(self.corpus),
+            # number of single word utterances
+            'nutterances_single_word': wlen.count(1),
+            # number of word tokens
+            'nword_tokens': sum(wlen),
+            # number of word types
+            'nword_types': len(self.unigram),
+            # number of word types with a frequency of 1 (hapax)
+            'nword_hapax': list(self.unigram.values()).count(1 / sum(wlen)),
+            # mean ratio of uniques words per chunk of 10 words
+            'mattr': sum(nuniques) / len(nuniques)
+        }
+
+        # average word length in number of phonemes
+        if 'phone' in self.separator.levels():
+            stats['awl'] = sum(
+                len(utt) for utt in self.tokens['phone']) / sum(wlen)
+
+        return stats
+
+    def top_frequency_tokens(self, level, n=None):
+        return collections.Counter(
+            (t for utt in self.tokens[level] for t in utt)).most_common(n)
+
+    def normalized_segmentation_entropy(self):
+        """Return the Normalized Segmentation Entropy computed on `text`
+
+        Token separators must be defined for phones and words.
+
+        Returns
+        -------
+        entropy : float
+            The estimated NSE in bits.
+
+        Raises
+        ------
 
 
-def stat_corpus(text, separator=default_separator):
-    """Return basis descriptive statistics of a text corpus
+        Notes
+        -----
+        As explained in [1]_ we are interested in the ambiguity generated
+        by the different possible parses that result from a
+        segmentation. In order to quantify this idea in general, we define
+        a Normalized Segmentation Entropy. To do this, we need to assign a
+        probability to every possible segmentation. To this end, we use a
+        unigram model where the probability of a lexical item is its
+        normalized frequency in the corpus and the probability of a parse
+        is the product of the probabilities of its terms. In order to
+        obtain a measure that does not depend on the utterance length, we
+        normalize by the number of possible boundaries in the
+        utterance. So for an utterance of length N, the Normalized
+        Segmentation Entropy (NSE) is computed using Shannon formula
+        (Shannon, 1948) as follows:
 
-    :param sequence(str) text: the input sequence to process, each
-      string in the sequence is an utterance
+        .. math::
 
-    :param Separator separator: token separation in the `text`
+            NSE = -\sum_i P_ilog_2(P_i) / (N-1),
 
-    :return: A pandas.DataFrame containing the number of tokens, types
-      and utterances in the input `text`
+        where :math:`P_i` is the probability of the word :math:`i` and
+        :math:`N` the number of phonemes in the text.
 
-    :todo: add average word length in number of syllables, average
-       word length in number of phonemes
+        .. [1] A. Fourtassi, B. Börschinger, M. Johnson and E. Dupoux,
+           "Whyisenglishsoeasytosegment". In Proceedings of the Fourth Annual
+           Workshop on Cognitive Modeling and Computational Linguistics
+           (pp. 1-10), 2013.
 
-    """
-    # force the text as a list if it is a generator
-    text = list(text)
+        """
+        # count the number of phones in the text
+        N = sum(len(utt) for utt in self.tokens['phone'])
 
-    list_of_words = []
-    dict_of_types = collections.Counter()
-    for line in text:
-        for word in separator.split(line, 'word'):
-            list_of_words.append(word)
-            dict_of_types.update(word)
+        # word lexicon with probabilities
+        P = self.unigram
 
-    df = pd.DataFrame(
-        index=['stat'],
-        columns=['number_tokens', 'number_types', 'number_utterances'])
+        # the probability of each word in the text
+        probs = (P[word] for utt in self.tokens['word'] for word in utt)
 
-    df.number_tokens = len(list_of_words)
-    df.number_types = len(dict_of_types)
-    df.number_utterances = len(text)
-
-    return df
-
-
-def top_frequency_tokens(text, n=1000, separator=default_separator,
-                         level='word'):
-    """Return the `n` most common tokens in `text`
-
-    :param list(str) text: the nput text split in utterances
-
-    :param int n: the most common tokens to return
-
-    :param Separator separator: tokens separation in `text`
-
-    :param str level: the token level to consider, must be 'word',
-        syllable' or 'phoneme'.
-
-    :return: a list of `n` (token, count) pairs of most common tokens in `text`
-
-    """
-    tokens = (token for line in text for token
-              in separator.split(line.strip(), level))
-    return collections.Counter(tokens).most_common(n)
+        # compute the entropy
+        return -1 * sum(p * log2(p) / (N - 1) for p in probs)
 
 
 @utils.CatchExceptions
@@ -79,19 +158,15 @@ def main():
     streamin, streamout, separator, log, args = utils.prepare_main(
         name='wordseg-stats',
         description=__doc__,
-        separator=default_separator)
+        separator=Separator())
 
-    # force the input to be a list (can be a sequence or a file buffer)
-    streamin = list(streamin)
-
-    top = top_frequency_tokens(streamin, separator=separator)
+    stats = CorpusStatistics(streamin, separator, log=log)
+    basics = stats.describe_corpus()
+    streamout.write(' '.join(basics.keys()) + ' nse\n')
     streamout.write(
-        '{} top frequency tokens:\n'.format(len(top))
-        + '\n'.join('{} {}'.format(t[0], t[1]) for t in top)
-        + '\n' * 2)
-
-    stat = stat_corpus(streamin, separator=separator)
-    streamout.write(stat.to_string() + '\n')
+        ' '.join(str(v) if isinstance(v, int) else '{0:0.4f}'.format(v)
+                 for v in basics.values()) + ' {0:0.4f}'.format(
+                         stats.normalized_segmentation_entropy()) + '\n')
 
 
 if __name__ == '__main__':